Add configurable memory based throttle implementation.
In order to avoid contesting the database rows and connection
to provide throttling, instead use a Redis based throttle system.

This is based on a token bucket based system.

It should be noted that this throttle may race slightly when under
very heavy load, but this is an acceptable compromise in the
implementation to keep things simple and small.

If you're planning on using this and expect it to remain 100%
accurate under very high rates, look into a different approach
such as redis sorted sets and obtaining set sizes.