Show deleted post thumbnails
Allow users to see and click on deleted posts without showing
the actual post preview. This roughly equates to how deleted
posts work on E621 at present.

This also resulted in some clean up and bug fixing surrounding
determining if a post should be visible or not. Previously this
was limited to user level(for ugrade incentive), safe mode, and
if the post was banned. Now it's only safe mode and if the post
is deleted.

Future work items:
 - Remove display deleted toggle and support code, as it is broken.