Test and bring tickets closer to completion
Add links to report items to all of the reportable areas of the
site.

Fix a lot of errors surrounding views and bad expectations from
the old codebase.

Change how dmails can be viewed to match the old logic. This prevents
users from creating links to their own dmails that can be shared with
moderators, and instead forces them to be reported.

Sets are still not implemented, so the reporting area for them has
not been tested or revised. That will happen at a later phase when
sets go in.