Add back previous post query builder until migration is finished
A lot of code relies on little bits and pieces of the old query builder,
either for utility functions for returning information in a specific way.

Until all of these code sites have been refactored, the old parser needs
to exist within the codebase for it to continue to operate.

Primarily this impacts various search functions around the site using
the numerical search helper functions.