Don't queue nil tag category update jobs
Tag category updates don't need to run for new tags, and attempting
to run them results in an exception because the tag does not have
a tag id at that point.

The tag counts for the post are still correct as the tag has been
saved and is in the transaction by the time the post system asks
for tag counts for each category, negating the need for the job.