Internalize vendored scripts and css
This avoids leaking usage data and also prevents us from having to
wait on the extremely slow public FA CDN.

These may be distributed to the static content CDN later, so
have integrity markers applied to them ahead of time.

spark-md5 may be removed and integrated directly into the bundle
when the new upload page is implemented, and is currently unused.