Fix regression in JS linking
This fixes a regression introduced in PR #75 where I mistakenly
removed the exports from the application pack. These were being
placed into the Danbooru namespace object as a library and referenced
by various items across the site.

This brings them back and converts the takedown and blip services
to use the namespace.