Replace google cloud with ElasticSearch
Most of this was related to use of BigQuery for post versions and
generating reports and bulk reverts, but elasticsearch can easily
handle returning 5k results in a few ms, so this isn't a big deal.

There is also some pubsub stuff which has been gone for a while now.
I don't remember what it was being used for, but it is likely reporting
items, so not an immediate priority.