Customize new users before save
The old before_create callback fired too late and it was almost
certain to overwrite changes to the model made elsewhere without
very careful orchestration. This moves customization closer
to instance creation and averts many potential problems.

The customize method still needs to take into account that models
can be customized during creation with hash params and check
for those changes for important parameters.

The better way to do this is probably to change the defaults in
the database, but it's less customizable.