Prevent auto tags from showing as user input
The -status:deleted and rating:s tags could be automatically added
by the system. Previously these were treated as user input and
shown as explicit search terms to the user.

This splits the automatic tags out from what the user entered so
that they don't appear in search boxes as user input.