Fix style of post_flags/events
Previously the tables on these pages had row color overrides with
static colors. These colors conflicted with the theme pretty
heavily, and were fixed.

During this process I noticed that it's common that pages want
the normal styling of tables that goes along with the striped class
but don't want them to be striped. So refactored the table classes
a bit to allow for a more generic table class and then optional
striping on top of it. The striped class extends the base class
right now because the base class didn't exist before this commit
so it's almost exclusively used in isolation instead of composed.