Fix email verification being a user level
Having a distinct level for user verification proved to be rather
problematic for code flow and introduced an extra level below
blocked. Since it makes the account unusable until activated
it made sense to use an extra field and stop messing with user
levels explicitly.

This should also make tests a lot easier to deal with because it
doesn't require any changing of the user level.