Fix bad error message on new artist page
The page was validating the model for default parameters but was
not checking if there were any values provided. Since the name
if a possible parameter that could be provided, it would say that
the name was invalid on initial visit.

To fix this validations are only run before page show if a name
is provided.