Fix result window on search
The default result window is 10k in ES, which is below our max
result count of 320 posts over 750 pages. Added 10k to the cap
just in case something comes up and we need more results.

I have been told that this doesn't cause problems in production
on Derpi, but we'll see. This can be lowered at runtime, so it
won't be a big deal if we have to roll it back down again.