Update date handling
E6 has a fairly complex date searching system that accepts several
natural language constructs. This ports some of those constructs
to the new search parser while leaving the "age:" metatag more or
less intact. This allows for such strange searches as "date:week"
to get all the posts that are newer than a week old.

This also updates the date metatag to accept many of the same
arguments as the age metatag takes. "ago" suffixes are not optional
and entirely ignored, but no longer invalidate the search.