Fix broken mobile gestures
When mobile gestures were enabled they would steal the normal
input from the browser. This prevented zooming in and out as well
as horizontal panning. Hammer didn't have a good configuration
system and so I swapped to using ZingTouch which is much more
simple, and offers better configuration. This allowed for easy
integration of gestures without stealing all input from the
browser.