[Deletion] Optimize post deletion slightly
- Don't move sets to parents unless moving favorites.
- Allow for a longer timeout on deletion.
- Move undelete into a transaction.
- Move actions that are not important out of the transaction.
  This allows for moving of sets to fail without making the post
  impossible to delete.

TODO: Move deletion back into a background job because it
can take a long time now that it's moving set items again?

Background job deletion was the cause of weird race conditions
and errors, but could prove to make moderator lives much much
easier. With a dozen sets it can take 30+ seconds to finish a post
deletion, which is much less than ideal.