[Util] Validate page numbers
Because it seems that **SOME** browsers have a bad habit of stuffing
garbage into the page parameter, add some additional validation to
the paginator to explicitly reject invalid page numbers.

Most notable cases seem to be copying the page title into the page
param, which then gets processed in an undefined way by other code
and throws an exception.