[Sources] Normalize pixiv source urls on save
Pixiv changed their url format(again?!) and so the alternate url
processor needed to be updated to account for these new urls.

In addition to this it was generating duplicate source urls, so
new code was needed to automatically remove old source urls that
would be duplicates of the new alternates that are generated.

I have decided to focus on the https://www.pixiv.net/artworks/id
sources for right now, as they automatically redirect to the /en/
version as needed based on region, and is more friendly to global
users of the site.