[Wiki] Fix aliases/implications not showing on missing tag
It's possible for aliases and implications to be present for a tag
without that tag existing in the database(usually due to aliases.)

If this happens, populate a new tag with the name of the wiki page
for the scope of that check so it can look up these otherwise
orphaned tag relationships.

This is hacky, but it means not having to rewrite all of the
surrounding logic for obtaining alias and implication lists by
a string instead of based on a tag instance.