[Post History] Fix for incorrect search method
With the 7.1.1 update to elasticsearch-model a regression was fixed
where the search method would be blindly deleted to elasticsearch.
Because our models(stupidly) always define the search method this
caused it to pull up all tag history instead of searching correctly.

This explicitly points at the elasticsearch search method and fixes
the problem.

Never figured I'd be bitten by a bug fix for a regression.