[UI] Fix multiple static navbar entries being active
Previously the controller was used exclusivley to check if a route
was active. Because static entries all use the same controller it was
possible for multiple navbar entries to be active at the same time
if you were on a static page.

This fixes that by first checking for route equality and falling back
to the site map later on