[Debug] Better logging of slow queries
Rails is VERY opaque about which query timed out when reviewing
backtraces for code that loads relations. Often it becomes difficult
to tell exactly which query was generated.

This interrogates the exceptions for these and pulls out the SQL
and bind params for further debugging. Automatic unwrapping of
rails template exceptions is done, as a majority of these seem
to be triggered in production by relations that are lazily loaded
through templates, or template code.

Also I cleaned up the backtrace display a little bit while I was
messing with this code. Usually it isn't relevant to see the 130
lines of rails relationship loader code. Full traces are still
saved so that in the event something happens inside a gem/framework
file it can still be easily located.