[Posts] Make post html generator detect status case insensitive
When using status:Deleted the search system would properly downcase
the value, but no results were shown. This is because the status was
being checked against in the post presenter and decorator in order to
prevent showing deleted thumbnails without circumstance.

This makes the matching on the status check case insensitive, which
fixes status:Deleted