[Thumbs] Fix thumbnail error when icc transform isn't possible
Newer versions of vips seem very picky about being able to match
the input icc profile to the input image. This fails for some
images that don't have the correct number of input bands.
This tries the image again without an input icc profile and
allows the processing to continue with only a warning.

Fixes black and white/grayscale png inputs as well as some other
incompatible input modes that don't have 3 bands and so can't be
treated as RGB.