[JS] Fix webpacker overriding babel config and setting loose
Loose does super dumb shit like converts spread operators to
[].concat() which is broken, incorrect, and not at all helpful.

Because this was forced in their base preset, it would break all
kinds of things that shouldn't have broken, and it would ignore the
project config explicitly.

This seems to fix the vast majority of the JS on the site, and the
generated code makes a lot more sense now.