[TagRelations] Fix searching with only one category
Sometimes the joined table is named "tags", sometimes "consequent_tag".
This depends on wether or not both antecedent and consequent tag where joined against.
Because you can only specifiy a string or symbol for ordering and not a nested hash
this will have to do.

Also remove the legacy params. Turns out I broke them a while ago,
but no one noticed.