[DText] Fix browser offering scroll in certain cases
Happens if the DText has newlines, the content has the update notice and
it's in a table. Can happens on the user feedback index.
There seems to be miniscule overflow happening for some reason. Just hide it.
I originally added this overflow here for zalgo text, which still works this way.