[Users] Update turkish welcome message
When I logged onto the website today and saw the welcome page on the
home screen, I noticed that the Turkish version of the splash had the word
"Karşılama" on it (I linked the screenshot). This doesn't mean "welcome" as
an exclamation, but rather the verb (as in "to welcome").  A better replacement
for this would be "Hoşgeldin" or "Hoşgeldiniz," depending on how formal you
want the greeting to be (the former being less formal). I hope this helps!