[Posts] Fix day/week/month not line-breaking on the popular page
This previously had whitespace that I unintentionally removed. This also has the unintentional sideeffect
of not breaking up individual prev/scale/next links which I see as an improvement.

Remove a margin override for the first entry, this wasn't properly centered for some reason.
Still not quite there but much better now.